(1)修改了 只生成一个Excel文件 (2)美化窗体

+++ Filter/src/main/java/gov/sc/file/WriteExcelFile.java
@@ -1,5 +1,7 @@
 package gov.sc.file;
 
+import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.List;
@@ -15,9 +17,9 @@ import org.apache.poi.xssf.usermodel.XSSFWorkbook;
  * 该类继承自File类
  * 
  * @author Kevin
- *
+ * 
  */
-public class WriteExcelFile{
+public class WriteExcelFile {
 
 	private String file;
 
@@ -52,26 +54,52 @@ public class WriteExcelFile{
 	 * @throws IOException
 	 */
 	@SuppressWarnings("resource")
-	public void write(List<String[]> list) throws IOException {
+	public void write(List<String[]> list, int number) throws IOException {
 		Workbook workbook;
-		if (file.endsWith(".xls")) {
-			workbook = new HSSFWorkbook();
-		} else {
-			workbook = new XSSFWorkbook();
-		}
-		Sheet sheet = workbook.createSheet("sheet1");
-		
-		for (int i = 0; i < list.size(); i++) {
-			String[] rowList = list.get(i);
-			Row row = sheet.createRow(i);
-			for (int j = 0; j < rowList.length; j++) {
-				Cell cell = row.createCell(j);
-				cell.setCellValue(rowList[j]);
+		File file1 = new File(file);
+		if (file1.exists()) {
+			
+			if (file.endsWith(".xls")) {
+				workbook = new HSSFWorkbook(new FileInputStream(file));
+			} else {
+				workbook = new XSSFWorkbook(new FileInputStream(file));
 			}
+			Sheet sheet = workbook.createSheet("sheet" + number);
+
+			for (int i = 0; i < list.size(); i++) {
+				String[] rowList = list.get(i);
+				Row row = sheet.createRow(i);
+				for (int j = 0; j < rowList.length; j++) {
+					Cell cell = row.createCell(j);
+					cell.setCellValue(rowList[j]);
+				}
+			}
+			FileOutputStream fout = new FileOutputStream(file);
+			workbook.write(fout);
+			fout.flush();
+			fout.close();
+
+		} else {
+
+			if (file.endsWith(".xls")) {
+				workbook = new HSSFWorkbook();
+			} else {
+				workbook = new XSSFWorkbook();
+			}
+			Sheet sheet = workbook.createSheet("sheet" + number);
+
+			for (int i = 0; i < list.size(); i++) {
+				String[] rowList = list.get(i);
+				Row row = sheet.createRow(i);
+				for (int j = 0; j < rowList.length; j++) {
+					Cell cell = row.createCell(j);
+					cell.setCellValue(rowList[j]);
+				}
+			}
+			FileOutputStream fout = new FileOutputStream(file);
+			workbook.write(fout);
+			fout.flush();
+			fout.close();
 		}
-		FileOutputStream fout = new FileOutputStream(file);
-		workbook.write(fout);
-		fout.flush();
-		fout.close();
 	}
 }
